Factor out bootstrap-related code.
