+++ environs/bootstrap.go
@@ -1,0 +1,64 @@
+// Copyright 2015 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package environs
+
+import (
+	"io"
+	"os"
+
+	"github.com/juju/juju/cloudconfig/instancecfg"
+	"github.com/juju/juju/constraints"
+	"github.com/juju/juju/tools"
+)
+
+// BootstrapParams holds the parameters for bootstrapping an environment.
+type BootstrapParams struct {
+	// Constraints are used to choose the initial instance specification,
+	// and will be stored in the new environment's state.
+	Constraints constraints.Value
+
+	// Placement, if non-empty, holds an environment-specific placement
+	// directive used to choose the initial instance.
+	Placement string
+
+	// AvailableTools is a collection of tools which the Bootstrap method
+	// may use to decide which architecture/series to instantiate.
+	AvailableTools tools.List
+
+	// ContainerBridgeName, if non-empty, overrides the default
+	// network bridge device to use for LXC and KVM containers. See
+	// also instancecfg.DefaultBridgeName.
+	ContainerBridgeName string
+}
+
+// BootstrapFinalizer is a function returned from Environ.Bootstrap.
+// The caller must pass a InstanceConfig with the Tools field set.
+type BootstrapFinalizer func(BootstrapContext, *instancecfg.InstanceConfig) error
+
+// BootstrapContext is an interface that is passed to
+// Environ.Bootstrap, providing a means of obtaining
+// information about and manipulating the context in which
+// it is being invoked.
+type BootstrapContext interface {
+	GetStdin() io.Reader
+	GetStdout() io.Writer
+	GetStderr() io.Writer
+	Infof(format string, params ...interface{})
+	Verbosef(format string, params ...interface{})
+
+	// InterruptNotify starts watching for interrupt signals
+	// on behalf of the caller, sending them to the supplied
+	// channel.
+	InterruptNotify(sig chan<- os.Signal)
+
+	// StopInterruptNotify undoes the effects of a previous
+	// call to InterruptNotify with the same channel. After
+	// StopInterruptNotify returns, no more signals will be
+	// delivered to the channel.
+	StopInterruptNotify(chan<- os.Signal)
+
+	// ShouldVerifyCredentials indicates whether the caller's cloud
+	// credentials should be verified.
+	ShouldVerifyCredentials() bool
+}
